Move CLI folder
