+++ cli/.npmignore
