+++ cli/src/testBucketBuilder.ts
@@ -1,10 +1,10 @@
-import IBMi from "@halcyontech/vscode-ibmi-types/api/IBMi";
-import { ApiUtils } from "./api/apiUtils";
-import { CCLVL, LogLevel, TestBucket } from "./api/types";
+import IBMi from "vscode-ibmi/src/api/IBMi";
+import { ApiUtils } from "../../api/apiUtils";
+import { CCLVL, LogLevel, TestBucket } from "../../api/types";
 import { TestOutputLogger } from "./loggers/testOutputLogger";
 import Parser from "vscode-rpgle/language/parser";
-import { IfsConfigHandler, LocalConfigHandler, QsysConfigHandler } from "./api/config";
-import { IFSFile } from "@halcyontech/vscode-ibmi-types/api/types";
+import { IfsConfigHandler, LocalConfigHandler, QsysConfigHandler } from "../../api/config";
+import { IFSFile } from "vscode-ibmi/src/api/types";
 import * as fs from 'fs';
 import * as path from 'path';
