+++ package.json
@@ -231,24 +231,20 @@
     "package": "webpack --mode production --devtool hidden-source-map",
     "pretest": "npm run compile && npm run lint",
     "lint": "eslint src --ext ts",
-    "preinstall": "cd src/cli && npm install && cd ../.."
+    "preinstall": "cd api && npm install && cd .."
   },
   "dependencies": {
-    "vscode-rpgle": "github:halcyon-tech/vscode-rpgle",
-    "ansi-colors": "^4.1.3",
     "compare-versions": "^6.1.1",
     "node-fetch": "^3.3.2",
     "octokit": "^4.1.2",
     "tmp": "^0.2.3",
-    "xml2js": "^0.6.2"
+    "vscode-ibmi": "github:halcyon-tech/vscode-ibmi",
+    "vscode-rpgle": "github:halcyon-tech/vscode-rpgle"
   },
   "devDependencies": {
-    "@halcyontech/vscode-ibmi-types": "^2.16.0",
-    "@types/mocha": "^10.0.7",
     "@types/node": "20.x",
     "@types/tmp": "^0.2.6",
     "@types/vscode": "^1.96.0",
-    "@types/xml2js": "^0.4.14",
     "@typescript-eslint/eslint-plugin": "^7.14.1",
     "@typescript-eslint/parser": "^7.11.0",
     "eslint": "^8.57.0",
