--- src/cli/.vscode/launch.json
@@ -1,56 +1,0 @@
-{
-	// Use IntelliSense to learn about possible attributes.
-	// Hover to view descriptions of existing attributes.
-	// For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-	"version": "0.2.0",
-	"configurations": [
-		{
-			"name": "Debug itest",
-			"type": "node",
-			"request": "launch",
-			"cwd": "${workspaceFolder:cli}",
-			"program": "${workspaceFolder:cli}/dist/index.js",
-			"sourceMaps": true,
-			"args": [
-				"--local-directory",
-				"../../examples/ibmi-company_system/",
-				"--ifs-directory",
-				"/home/$USER/builds/ibmi-company_system",
-				"--library-list",
-				"RPGUNIT",
-				"QDEVTOOLS",
-				"--summary-report",
-				"--save-test-output",
-				"--save-test-result",
-				"--save-command-output"
-			],
-			"preLaunchTask": {
-				"type": "npm",
-				"script": "webpack:dev"
-			},
-			"env": {
-				"IBMI_HOST": "${input:ibmiHost}",
-				"IBMI_USER": "${input:ibmiUser}",
-				"IBMI_PASSWORD": "${input:ibmiPassword}",
-			}
-		}
-	],
-	"inputs": [
-		{
-			"id": "ibmiHost",
-			"type": "promptString",
-			"description": "Enter the IBMI_HOST"
-		},
-		{
-			"id": "ibmiUser",
-			"type": "promptString",
-			"description": "Enter the IBMI_USER"
-		},
-		{
-			"id": "ibmiPassword",
-			"type": "promptString",
-			"description": "Enter the IBMI_PASSWORD",
-			"password": true
-		}
-	]
-}
